fourth commit about day 9
add dictionary_in_list.py : You are going to write a program that adds to a travel_log. You can see a travel_log which is a List that contains 2 Dictionaries.
Write a function that will work with the following line of code on line 21 to add the entry for Russia to the travel_log.

fix nested_dictionary.py : Nesting a list in a Dictionary