The 47th commit
Create make_sketch.py code. This code make Etch-a-Sketch.
The new function I learned to write this code today is clear() and home() function.

clear() is Delete the turtle's drawings from the screen.
home() is Move turtle to the origin.

And onkey() function without arguments cannot be received.
So, onkey(fun, key) => 'fun' is a function with no arguments or None